split code and add unit test
